% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GelbeKarten_Spielverlauf_Mannschaft.R
\name{GelbeKarten_Spielminute_Mannschaft}
\alias{GelbeKarten_Spielminute_Mannschaft}
\title{Title}
\usage{
GelbeKarten_Spielminute_Mannschaft(Mannschaft, Data)
}
\arguments{
\item{Data}{}
}
\description{
Title
}
