% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GelbeKarten_Spielverlauf_Spieler.R
\name{GelbeKarten_Spielverlauf_Spieler}
\alias{GelbeKarten_Spielverlauf_Spieler}
\title{Title}
\usage{
GelbeKarten_Spielverlauf_Spieler(Name, Data)
}
\arguments{
\item{Data}{}
}
\description{
Title
}
