% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape_SIS.R
\name{Scrape_SIS}
\alias{Scrape_SIS}
\title{Diese Funktion tut etwas....}
\usage{
Scrape_SIS(liga = "001519505501509501000000000000000002007")
}
\arguments{
\item{liga}{die liga ID als string}
}
\value{
dunno something
}
\description{
Diese Funktion tut etwas....
}
\examples{
df <- Scrape_SIS("ABC")
}
