% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tore_Spielverlauf_Mannschaft.R
\name{Tore_Spielminute_Mannschaft}
\alias{Tore_Spielminute_Mannschaft}
\title{Title}
\usage{
Tore_Spielminute_Mannschaft(Mannschaft, Data)
}
\arguments{
\item{Data}{}
}
\description{
Title
}
