% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeitstrafen_Spielverlauf_Mannschaft.R
\name{Zeitstrafen_Spielminute_Mannschaft}
\alias{Zeitstrafen_Spielminute_Mannschaft}
\title{Title}
\usage{
Zeitstrafen_Spielminute_Mannschaft(Mannschaft, Data)
}
\arguments{
\item{Data}{}
}
\description{
Title
}
