% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeitrafen_Spielverlauf_Spieler.R
\name{Zeitstrafen_Spielverlauf_Spieler}
\alias{Zeitstrafen_Spielverlauf_Spieler}
\title{Title}
\usage{
Zeitstrafen_Spielverlauf_Spieler(Name, Data)
}
\arguments{
\item{Data}{}
}
\description{
Title
}
